@@ -26,12 +26,11 @@ class CachedNetworkImageProvider
     this.cacheManager,
     this.cacheKey,
     //ignore: avoid_unused_constructor_parameters
-    ImageRenderMethodForWeb imageRenderMethodForWeb,
-  })  : assert(url != null),
-        assert(scale != null);
+    ImageRenderMethodForWeb? imageRenderMethodForWeb,
+  });
 
   @override
-  final BaseCacheManager cacheManager;
+  final BaseCacheManager? cacheManager;
 
   /// Web url of the image to load
   @override
@@ -39,7 +38,7 @@ class CachedNetworkImageProvider
 
   /// Cache key of the image to cache
   @override
-  final String cacheKey;
+  final String? cacheKey;
 
   /// Scale of the image
   @override
@@ -47,17 +46,17 @@ class CachedNetworkImageProvider
 
   /// Listener to be called when images fails to load.
   @override
-  final image_provider.ErrorListener errorListener;
+  final image_provider.ErrorListener? errorListener;
 
   /// Set headers for the image provider, for example for authentication
   @override
-  final Map<String, String> headers;
+  final Map<String, String>? headers;
 
   @override
-  final int maxHeight;
+  final int? maxHeight;
 
   @override
-  final int maxWidth;
+  final int? maxWidth;
 
   @override
   Future<CachedNetworkImageProvider> obtainKey(
@@ -84,7 +83,7 @@ class CachedNetworkImageProvider
   }
 
   Stream<ui.Codec> _loadAsync(
-    CachedNetworkImageProvider key,
+    image_provider.CachedNetworkImageProvider key,
     StreamController<ImageChunkEvent> chunkEvents,
     DecoderCallback decode,
   ) async* {
@@ -126,7 +125,7 @@ class CachedNetworkImageProvider
       // have had a chance to track the key in the cache at all.
       // Schedule a microtask to give the cache a chance to add the key.
       scheduleMicrotask(() {
-        PaintingBinding.instance.imageCache.evict(key);
+        PaintingBinding.instance?.imageCache?.evict(key);
       });
 
       errorListener?.call();
